<MegaAVR> Removed redundancy of platform resource
- Removed weak functions
- Removed repeated code from platform folders
